kr3/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
